move PathLength class to new file to avoid circular imports
